requirements.txt
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
astor
asttokens
async-lru
attrs
babel
beautifulsoup4
bleach
certifi
cffi
charset-normalizer
colorama
comm
contourpy
cycler
Cython
debugpy
decorator
defusedxml
distro
entrypoints
et-xmlfile
exceptiongroup
executing
fastjsonschema
fire
fonttools
fqdn
fuzzywuzzy
h11
httpcore
httpx
idna
imageio
imgaug
ipykernel
ipython
isoduration
jedi
Jinja2
jiter
joblib
json5
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
kiwisolver
lazy_loader
Levenshtein
lmdb
lxml
MarkupSafe
matplotlib
matplotlib-inline
mistune
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numpy
openai
opencv-contrib-python
opencv-python
openpyxl
opt-einsum
outcome
overrides
packaging
paddleocr
paddlepaddle
pandas
pandocfilters
parso
pep517
pillow
platformdirs
prometheus_client
prompt_toolkit
protobuf
pure-eval
pyclipper
pycparser
pydantic
pydantic_core
Pygments
pyparsing
PySocks
pytesseract
python-dateutil
python-docx
python-dotenv
python-json-logger
python-Levenshtein
pytz
pywin32
pywinpty
PyYAML
pyzmq
rapidfuzz
referencing
requests
rfc3339-validator
rfc3986-validator
rpds-py
scikit-image
scikit-learn
scipy
selenium
Send2Trash
shapely
six
sniffio
sortedcontainers
soupsieve
stack-data
termcolor
terminado
threadpoolctl
tifffile
tinycss2
tomli
tornado
tqdm
traitlets
trio
trio-websocket
typing_extensions
tzdata
uri-template
urllib3
wcwidth
webcolors
webdriver-manager
webencodings
websocket-client
wsproto
xlrd